requirements.txt
# Core Python Dependencies
fastapi
uvicorn
pydantic

# LangChain & Embeddings
langchain==0.3.20
langchain-community==0.3.19
langchain-huggingface>=0.0.1
langchain-text-splitters==0.3.6


# LLM
llama-cpp-python==0.3.8 

# CrewAI
crewai==0.108.0


# Vector Search & Text
faiss-cpu==1.10.0
sentence-transformers==3.4.1

accelerate
numpy

# Image Processing
pillow
pytesseract
matplotlib
opencv-python-headless

# OCR Language Support
tesseract
# Note: Tesseract must also be installed on the system (already handled in Dockerfile via `apt-get install`)

# Fuzzy Matching
fuzzywuzzy
python-Levenshtein
thefuzz

# PDF/Text Parsing
python-docx
PyMuPDF
pdfplumber

# Excel Handling
openpyxl
pandas

# Logging & Utilities
loguru
requests
tqdm
nltk==3.8.1
autocorrect


# File Type Handling (Optional but helpful)

#Core Python Dependencies
# fastapi
# uvicorn
# pydantic

# # LangChain & CrewAI
# langchain==0.3.20
# langchain-community==0.3.19
# langchain-core==0.3.41
# langchain-huggingface>=0.0.1
# langchain-text-splitters==0.3.6
# crewai==0.108.0

# # Embeddings & Vector Search
# faiss-cpu==1.10.0
# sentence-transformers==3.4.1
# transformers==4.49.0
# llama-cpp-python==0.2.60
# numpy
# pandas

# # OCR and Image Processing
# pillow
# pytesseract
# opencv-python-headless
# matplotlib

# # PDF/Text Parsing
# PyMuPDF  # for PDFs
# pdfplumber  # if you really use it (optional)
# python-docx  # only if you're parsing .docx files

# # Excel
# openpyxl

# # Fuzzy Matching (Pick only ONE of the below)
# thefuzz  # actively maintained
# fuzzywuzzy  # ❌ redundant
# python-Levenshtein  # ❌ not required unless performance-critical

# # Utilities
# loguru
# requests
# tqdm